wrap up on the clog no longer being inherited

+++ include/Opensea_Parser_Helper.h
@@ -144,7 +144,7 @@ namespace opensea_parser {
 	{
 		if (length != 0)
 		{
-			for (int i = 0; i < (sizeof(pageCodes) / sizeof(*pageCodes)); i++)
+			for (size_t i = 0; i < (sizeof(pageCodes) / sizeof(*pageCodes)); i++)
 			{
 				if (pageCodes[i] == code)
 				{
